[MOD] : TicketViewHolder 분리 #130
