+++ app/src/main/java/com/indipage/presentation/ticket/ticket/viewholder/TicketViewHolder.kt
@@ -1,0 +1,14 @@
+package com.indipage.presentation.ticket.ticket.viewholder
+
+import androidx.recyclerview.widget.RecyclerView
+import com.indipage.databinding.ItemTicketTicketBinding
+import com.indipage.domain.entity.Ticket
+
+class TicketViewHolder(
+    private val binding: ItemTicketTicketBinding
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(data: Ticket) {
+        binding.ticket = data
+        binding.executePendingBindings()
+    }
+}
